numberplate/requirement.txt
mysql-connector-python
easyocr
opencv-python